Started working on organizer event view, waiting list, and drawing attendees
