+++ app/src/main/AndroidManifest.xml
@@ -58,6 +58,10 @@
             android:exported="true">
         </activity>
         <activity
+            android:name=".views.organizer.OrganizerEvent"
+            android:exported="true">
+        </activity>
+        <activity
             android:name=".views.profiles.OrganizerProfileActivity"
             android:exported="true">
         </activity>
