+++ app/src/main/java/com/example/queueup/viewmodels/OrganizerEventArrayAdapter.java
@@ -1,15 +1,31 @@
 package com.example.queueup.viewmodels;
 
-import android.content.Context;
-import android.view.View;
 
+import android.content.Context;
+import android.content.Intent;
+import android.os.Build;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.queueup.MainActivity;
 import com.example.queueup.R;
 import com.example.queueup.models.Event;
 import com.example.queueup.views.admin.AdminEventFragment;
-import com.example.queueup.views.attendee.AttendeeEventFragment;
 
+import com.example.queueup.views.organizer.OrganizerCreateEvent;
+import com.example.queueup.views.organizer.OrganizerEvent;
+import com.example.queueup.views.organizer.OrganizerHome;
+
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
 public class OrganizerEventArrayAdapter extends EventArrayAdapter {
@@ -19,11 +35,8 @@ public class OrganizerEventArrayAdapter extends EventArrayAdapter {
 
     protected View.OnClickListener onClickListener(View view) {
         return (v) -> {
-            AppCompatActivity activity = (AppCompatActivity) view.getContext();
-            activity.getSupportFragmentManager().beginTransaction()
-                    .replace(R.id.organizer_activity_fragment, AdminEventFragment.class, null)
-                    .addToBackStack(null)
-                    .commit();
+            Intent intent = new Intent(view.getContext(), OrganizerEvent.class);
+            view.getContext().startActivity(intent);
         };
     }
