+++ app/src/main/java/com/example/queueup/views/organizer/OrganizerWaitingListFragment.java
@@ -1,0 +1,34 @@
+package com.example.queueup.views.organizer;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ListView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
+import com.example.queueup.R;
+import com.example.queueup.models.User;
+import com.example.queueup.viewmodels.UsersArrayAdapter;
+
+import java.util.ArrayList;
+
+public class OrganizerWaitingListFragment extends Fragment {
+    public OrganizerWaitingListFragment() {
+        super(R.layout.organizer_waiting_list_fragment);
+    }
+    private ArrayList<User> dataList;
+    private ListView userList;
+    private UsersArrayAdapter usersAdapter;
+    @Override
+    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
+        User e = new User("e","e","e","e","e","e");
+        dataList = new ArrayList<User>();
+        dataList.add(e);
+
+        userList = getView().findViewById(R.id.event_waiting_list);
+        usersAdapter = new UsersArrayAdapter(view.getContext(), dataList);
+        userList.setAdapter(usersAdapter);
+    }
+
+}
