+++ app/src/main/res/drawable/draw.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="86dp"
+    android:height="80dp"
+    android:viewportWidth="86"
+    android:viewportHeight="80">
+  <path
+      android:pathData="M0,0h86v80h-86z"
+      android:fillColor="#ffffff"/>
+</vector>
