+++ app/src/main/res/layout/attendee_event_activity.xml
