+++ app/src/main/res/layout/organizer_event_navigation_bar.xml
@@ -17,4 +17,4 @@
     app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintLeft_toLeftOf="parent"
     app:layout_constraintRight_toRightOf="parent"
-    app:menu="@menu/organizer_bottom_nav_menu"/>
+    app:menu="@menu/organizer_event_nav_menu"/>
