+++ app/src/main/res/menu/organizer_bottom_nav_menu.xml
@@ -7,7 +7,7 @@
     <item
         android:id="@+id/nav_QRCode"
         android:icon="@drawable/ic_nav_qrcode"
-        android:title="Users" />
+        android:title="QR Code" />
     <item
         android:id="@+id/nav_profile"
         android:icon="@drawable/ic_nav_profile"
